Split Queue and QueueManager proxies
In order to clarify when commands are going to the queue and when commands are
going to the QueueManager, we split the proxies in two. Upon calling
Klass.queue, the QueueProxy is returned, which is duck-typed into an
ActorProxy for a `worker_class` instance.

However, to access the QueueManager (for things like resizing and mailbox
inspection), you can use the special `#__manager__` method, which brings up
the ActorProxy for the QueueManager instance itself.