Add Delegator for true Queue-like functionality
Delegator exposes the #perform_with method which accepts a delegatee class and
method arguments. Delegator will create an instance of the delegatee and call
the #perform method of the delegatee, providing the method arguments.

This allows us to have a single pool of resources which perform multiple
responsibilities, multiplexing various tasks together.